Rename tools/pkg to tools/internal to make packages private
