+++ tools/internal/regsync/config.go
